requirements.txt
Pillow
pyperclip
pytesseract